Replaced RFloat by RDouble
git-svn-id: file:///home/svn/shoola-omero/trunk@6099 6f2cb1de-eb0d-0410-b157-e593188b5901

+++ SRC/org/openmicroscopy/shoola/agents/imviewer/browser/BrowserModel.java
@@ -634,7 +634,7 @@ class BrowserModel
      * 
      * @return See above.
      */
-    float getPixelsSizeX() { return parent.getPixelsSizeX(); }
+    double getPixelsSizeX() { return parent.getPixelsSizeX(); }
     
     /**
      * Returns <code>true</code> if the unit bar is painted on top of 
@@ -828,14 +828,14 @@ class BrowserModel
      * 
      * @return See above.
      */
-    float getPixelsSizeY() { return parent.getPixelsSizeY(); }
+	double getPixelsSizeY() { return parent.getPixelsSizeY(); }
 
     /**
      * Returns the size in microns of a pixel along the Y-axis.
      * 
      * @return See above.
      */
-    float getPixelsSizeZ() { return parent.getPixelsSizeZ(); }
+	double getPixelsSizeZ() { return parent.getPixelsSizeZ(); }
    
     /** 
      * Returns the number of column for the grid.
