+++ SRC/org/openmicroscopy/shoola/env/rnd/RenderingControlProxy.java
@@ -1065,7 +1065,7 @@ class RenderingControlProxy
      * Implemented as specified by {@link RenderingControl}. 
      * @see RenderingControl#getPixelsPhysicalSizeX()
      */
-    public float getPixelsPhysicalSizeX()
+    public double getPixelsPhysicalSizeX()
     {
         if (pixs.getPhysicalSizeX() == null) return 1;
         return pixs.getPhysicalSizeX().getValue();
@@ -1075,7 +1075,7 @@ class RenderingControlProxy
      * Implemented as specified by {@link RenderingControl}. 
      * @see RenderingControl#getPixelsPhysicalSizeY()
      */
-    public float getPixelsPhysicalSizeY()
+    public double getPixelsPhysicalSizeY()
     {
         if (pixs.getPhysicalSizeY() == null) return 1;
         return pixs.getPhysicalSizeY().getValue();
@@ -1085,7 +1085,7 @@ class RenderingControlProxy
      * Implemented as specified by {@link RenderingControl}. 
      * @see RenderingControl#getPixelsPhysicalSizeZ()
      */
-    public float getPixelsPhysicalSizeZ()
+    public double getPixelsPhysicalSizeZ()
     {
         if (pixs.getPhysicalSizeZ() == null) return 1;
         return pixs.getPhysicalSizeZ().getValue();
