+++ SRC/org/openmicroscopy/shoola/util/ui/lens/ZoomWindow.java
@@ -100,7 +100,7 @@ class ZoomWindow
 	 * @param x mapping in x axis.
 	 * @param y mapping in y axis.
 	 */
-	void setXYPixelMicron(float x, float y) 
+	void setXYPixelMicron(double x, double y) 
 	{
 		zoomWindowUI.setXYPixelMicron(x, y);
 	}
